extract mailer to its own gradle project
