+++ server/Dockerfile
@@ -5,7 +5,7 @@ FROM gradle:7.4.2-jdk17 AS base
 
 WORKDIR /home/gradle/hub
 
-RUN mkdir catalog curatorship discovery hub-app hub-cli persistence techtransfer \
+RUN mkdir catalog curatorship discovery hub-app hub-cli mailer persistence techtransfer \
     && chown gradle:gradle . -R
 
 USER gradle
@@ -18,6 +18,7 @@ COPY --chown=gradle:gradle  curatorship/build.gradle.kts     curatorship/gradle.
 COPY --chown=gradle:gradle  discovery/build.gradle.kts       discovery/gradle.properties        ./discovery/
 COPY --chown=gradle:gradle  hub-app/build.gradle.kts         hub-app/gradle.properties          ./hub-app/
 COPY --chown=gradle:gradle  hub-cli/build.gradle.kts         hub-cli/gradle.properties          ./hub-cli/
+COPY --chown=gradle:gradle  mailer/build.gradle.kts          mailer/gradle.properties           ./mailer/
 COPY --chown=gradle:gradle  persistence/build.gradle.kts     persistence/gradle.properties      ./persistence/
 COPY --chown=gradle:gradle  techtransfer/build.gradle.kts    techtransfer/gradle.properties     ./techtransfer/
