+++ server/curatorship/build.gradle.kts
@@ -4,7 +4,6 @@ val logback_version: String by project
 val mockk_version: String by project
 val ktor_version : String by project
 val typesafe_config_version: String by project
-val javax_mail_version: String by project
 
 plugins {
     kotlin("jvm")
@@ -21,6 +20,8 @@ repositories {
 }
 
 dependencies {
+    implementation(project(":mailer"))
+
     implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:$kotlinx_serialization_version")
     implementation("ch.qos.logback:logback-classic:$logback_version")
     implementation("io.ktor:ktor-client-core:$ktor_version")
@@ -28,7 +29,6 @@ dependencies {
     implementation("io.ktor:ktor-client-content-negotiation:$ktor_version")
     implementation("io.ktor:ktor-serialization-kotlinx-json:$ktor_version")
     implementation("com.typesafe:config:$typesafe_config_version")
-    implementation("javax.mail:mail:$javax_mail_version")
     testImplementation("org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version")
     testImplementation("io.mockk:mockk:$mockk_version")
 }
