+++ server/mailer/gradle.properties
@@ -1,7 +1,5 @@
 kotlin_version=1.7.10
 logback_version=1.2.3
 mockk_version=1.12.4
-kotlinx_serialization_version=1.3.3
-ktor_version=2.1.0
-typesafe_config_version=1.4.2
+javax_mail_version=1.5.0-b01
 kotlin.code.style=official
