+++ server/settings.gradle.kts
@@ -1,3 +1,3 @@
 rootProject.name = "br.usp.inovacao.hubusp.server"
 
-include("catalog", "curatorship", "discovery", "techtransfer", "persistence", "hub-cli", "hub-app")
+include("mailer", "catalog", "curatorship", "discovery", "techtransfer", "persistence", "hub-cli", "hub-app")
